Change linetext creation mode (draft)
